+++ src/ultra/hardcore.clj
@@ -1,5 +1,9 @@
 (ns ultra.hardcore
-  (:require [robert.hooke :refer [add-hook]]
+  "See what I did with the ns name?"
+  (:require [clojure.edn :as edn]
+            [clojure.java.io :as io]
+            [robert.hooke :refer [add-hook]]
+            [ultra.colorscheme :refer [set-colorscheme]]
             [ultra.repl :as repl]
             [ultra.stacktrace :as stacktrace]
             [ultra.test]))
@@ -7,11 +11,12 @@
 (def configured? (atom {}))
 
 (defn run-configuration
-  []
+  [opts]
+  (set-colorscheme opts)
   (repl/configure-repl!)
   (stacktrace/configure-stacktraces!)
   (ultra.test/activate!))
 
-(defn configure! []
+(defn configure! [opts]
   (if (empty? @configured?)
-    (swap! configured? assoc :config (run-configuration))))
+    (swap! configured? assoc :config (run-configuration opts))))